feat: Hubot Brain with MongoDB as datasource.
Hubot Brain by default uses an JS Object as source of data
then all data is saved on memory.

There are plugins that help to persist that object on a
DB after N secodns, but doesn't remove the dependency of
memory.

So we overwrite the brain, especifically the set & get methods
to save or query data directly from the DB.

Before these changes, the Brain was return data directly,
now We return a Promise, then is necesary update a little
your current implementation of methods: set & get.